fix  (core): Prepare for upcoming NettyHttpServer
