+++ java/dev/enola/web/WebServer.java
@@ -28,6 +28,9 @@ import java.net.InetSocketAddress;
  */
 public interface WebServer extends AutoCloseable {
 
+    // TODO Rename WebServer to HttpServer
+    // TODO Move package dev.enola.web to dev.enola.common.net.http
+
     void register(String path, WebHandler h);
 
     void start();
