+++ java/dev/enola/web/netty/NettyHttpServerTest.java
@@ -1,7 +1,7 @@
 /*
  * SPDX-License-Identifier: Apache-2.0
  *
- * Copyright 2023-2024 The Enola <https://enola.dev> Authors
+ * Copyright 2024 The Enola <https://enola.dev> Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,22 +15,29 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.enola.web;
+package dev.enola.web.netty;
 
-import com.google.common.util.concurrent.ListenableFuture;
+import dev.enola.web.WebServer;
+import dev.enola.web.testlib.WebServerTestAbstract;
 
-import dev.enola.common.io.resource.ReadableResource;
+import org.junit.Ignore;
+import org.junit.Test;
 
-import java.net.URI;
+import java.io.IOException;
 
-/**
- * Handler for Web Request.
- *
- * @see WebServer
- */
-public interface WebHandler {
+public class NettyHttpServerTest extends WebServerTestAbstract {
 
-    // TODO Rename WebHandler to dev.enola.common.net.http.HttpHandler
+    @Override
+    protected WebServer create() throws IOException {
+        return new NettyHttpServer();
+    }
 
-    ListenableFuture<ReadableResource> get(URI uri);
+    @Test
+    @Ignore // TODO
+    public void testServer() throws IOException {
+        super.testServer();
+    }
+
+    @Test
+    public void empty() {}
 }
